requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cssselect2==0.8.0
dj-database-url==2.3.0
Django==5.2
djangorestframework==3.16.0
fonttools==4.57.0
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
requests==2.32.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
weasyprint==65.0
webencodings==0.5.1
zopfli==0.2.3.post1
